Add files via upload
Using the python programming language, I created an anomaly detection procedure by following the five steps. These are bellow:
Step-1: Create a 2-d array of dimensions 100x1000. The elements should contain random values between 0 and 10. Name the array \normal" and save it to a binary le (i.e. non text le) with the name normal.bin; Repeat the same process for another array named \abnormal". The array should contain random values between 5 and 15. Save the array in a le called abnormal.bin.
Step-2: Load the saved data into two arrays (different variables).
Step-3: Manipulate these two arrays so that you create two new sets: (a) one array named \training" that contains 90% of the data of the normal array, (b) one that array named \test" that contains the other 10% of the data of the normal array and 10% of the abnormal array.
Step-4: For the first element (row) in the \training" set calculate the Euclidean distance of that element to all the rest of the elements in that set (each other row). Retrieve the top five distances and sum them. This is the dissimilarity score of that element to rest of the training set. Repeat the process for all other elements. The scores from this process should be stored in a vector named \baseline".
Step-5: For the first element of the \test" set, calculate the euclidean distance of that element to all elements contained in the \training" set. Retrieve the
top 5 distances and sum them. That is the dissimilarity score for that element. If the score is between the min-max values of the \baseline" 
flag that element as normal else  flag it as abnormal. Repeat the process for each element in the test set and print the algorithm's predictions.